DRTII-1567 Refactor to split dao & tables. Move table to package
