+++ jvm/src/main/scala/uk/gov/homeoffice/drt/db/dao/PassengersHourlyDao.scala
@@ -3,7 +3,7 @@ package uk.gov.homeoffice.drt.db.dao
 import slick.dbio.Effect
 import slick.sql.FixedSqlAction
 import uk.gov.homeoffice.drt.db.Db.slickProfile.api._
-import uk.gov.homeoffice.drt.db.{PassengersHourlyRow, PassengersHourlyTable}
+import uk.gov.homeoffice.drt.db.tables.{PassengersHourlyRow, PassengersHourlyTable}
 import uk.gov.homeoffice.drt.ports.PortCode
 import uk.gov.homeoffice.drt.ports.Queues.Queue
 import uk.gov.homeoffice.drt.ports.Terminals.Terminal
