+++ jvm/src/main/scala/uk/gov/homeoffice/drt/db/serialisers/CapacityHourlySerialiser.scala
@@ -1,6 +1,6 @@
 package uk.gov.homeoffice.drt.db.serialisers
 
-import uk.gov.homeoffice.drt.db.{CapacityHourly, CapacityHourlyRow}
+import uk.gov.homeoffice.drt.db.tables.{CapacityHourly, CapacityHourlyRow}
 import uk.gov.homeoffice.drt.ports.PortCode
 import uk.gov.homeoffice.drt.ports.Terminals.Terminal
 import uk.gov.homeoffice.drt.time.UtcDate
