+++ jvm/src/test/scala/uk/gov/homeoffice/drt/db/ABFeatureDaoSpec.scala
@@ -7,6 +7,7 @@ import slick.dbio.DBIO
 import scala.concurrent.Await
 import scala.concurrent.duration.DurationInt
 import slick.jdbc.PostgresProfile.api._
+import uk.gov.homeoffice.drt.db.tables.ABFeatureRow
 
 import java.sql.Timestamp
 import java.time.Instant
