+++ jvm/src/test/scala/uk/gov/homeoffice/drt/db/serialisers/CapacityHourlySerialiserTest.scala
@@ -2,7 +2,7 @@ package uk.gov.homeoffice.drt.db.serialisers
 
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
-import uk.gov.homeoffice.drt.db.{CapacityHourly, CapacityHourlyRow}
+import uk.gov.homeoffice.drt.db.tables.{CapacityHourly, CapacityHourlyRow}
 import uk.gov.homeoffice.drt.ports.PortCode
 import uk.gov.homeoffice.drt.ports.Terminals.Terminal
 import uk.gov.homeoffice.drt.time.UtcDate
