+++ jvm/src/test/scala/uk/gov/homeoffice/drt/db/serialisers/PassengersHourlySerialiserTest.scala
@@ -2,7 +2,7 @@ package uk.gov.homeoffice.drt.db.serialisers
 
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
-import uk.gov.homeoffice.drt.db.{PassengersHourly, PassengersHourlyRow}
+import uk.gov.homeoffice.drt.db.tables.{PassengersHourly, PassengersHourlyRow}
 import uk.gov.homeoffice.drt.ports.PortCode
 import uk.gov.homeoffice.drt.ports.Queues.Queue
 import uk.gov.homeoffice.drt.ports.Terminals.Terminal
